+++ tests/test_utilities/test_csvjson.py
@@ -30,6 +30,10 @@ class TestCSVJSON(CSVKitTestCase, EmptyFileTests):
         js = json.loads(self.get_output(['examples/sniff_limit.csv']))
         self.assertDictEqual(js[0], {'a': True, 'c': 3.0, 'b': 2.0})
 
+    def test_no_header_row(self):
+        js = json.loads(self.get_output(['--no-header-row', 'examples/no_header_row.csv']))
+        self.assertDictEqual(js[0], {'a': True, 'c': 3.0, 'b': 2.0})
+
     def test_no_inference(self):
         js = json.loads(self.get_output(['--no-inference', 'examples/dummy.csv']))
         self.assertDictEqual(js[0], {'a': '1', 'c': '3', 'b': '2'})